Traktor: Various GCC fixes.
[git-p4: depot-paths = "//depot/traktor/": change = 837]

+++ code/Flash/AccDisplayRenderer.cpp
@@ -256,7 +256,7 @@ void AccDisplayRenderer::end()
 #if defined(_DEBUG)
 	static int s_frame = 0;
 	if (s_frame++ % 100 == 0)
-		log::debug << L"Shape cache size " << m_shapeCache.size() << Endl;
+		log::debug << L"Shape cache size " << uint32_t(m_shapeCache.size()) << Endl;
 #endif
 }
